create `base` module
