+++ rpc/types/src/base/access_request_base.rs
@@ -8,21 +8,18 @@ use crate::Status;
 //---------------------------------------------------------------------------------------------------- ResponseBase
 /// TODO
 ///
-/// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/rpc/core_rpc_server_commands_defs.h#L101-L112>.
+/// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/rpc/core_rpc_server_commands_defs.h#L114-L122>.
 #[derive(
     Copy, Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize,
 )]
-pub struct ResponseBase {
+pub struct AccessRequestBase {
     /// TODO
-    status: Status,
-    /// TODO
-    untrusted: bool,
+    pub client: String,
 }
 
 epee_encoding::epee_object! {
-    ResponseBase,
-    status: Status,
-    untrusted: bool,
+    AccessRequestBase,
+    client: String,
 }
 
 //---------------------------------------------------------------------------------------------------- Tests
