+++ rpc/types/src/base/empty_response_base.rs
