+++ rpc/types/src/base/response_base.rs
@@ -8,18 +8,21 @@ use crate::Status;
 //---------------------------------------------------------------------------------------------------- ResponseBase
 /// TODO
 ///
-/// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/rpc/core_rpc_server_commands_defs.h#L114-L122>.
+/// <https://github.com/monero-project/monero/blob/cc73fe71162d564ffda8e549b79a350bca53c454/src/rpc/core_rpc_server_commands_defs.h#L101-L112>.
 #[derive(
     Copy, Clone, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize,
 )]
-pub struct AccessRequestBase {
+pub struct ResponseBase {
     /// TODO
-    client: String,
+    pub status: Status,
+    /// TODO
+    pub untrusted: bool,
 }
 
 epee_encoding::epee_object! {
-    AccessRequestBase,
-    client: String,
+    ResponseBase,
+    status: Status,
+    untrusted: bool,
 }
 
 //---------------------------------------------------------------------------------------------------- Tests
