+++ rpc/types/src/data/mod.rs
@@ -2,14 +2,6 @@
 //!
 //! TODO
 
-mod access_response_base;
 mod binary_string;
-mod empty_request_base;
-mod empty_response_base;
-mod response_base;
 
-pub use access_response_base::AccessResponseBase;
 pub use binary_string::BinaryString;
-pub use empty_request_base::EmptyRequestBase;
-pub use empty_response_base::EmptyResponseBase;
-pub use response_base::ResponseBase;
